fix linux build
