+++ ApprovalTests_SwiftTests/Scrubbers/ScrubbersTests.swift
@@ -1,7 +1,7 @@
-#if os(OSX)
-    import ApprovalTests_Swift
-#elseif os(iOS)
-    import ApprovalTests_iOS
+#if os(iOS)
+  import ApprovalTests_iOS
+#else
+  import ApprovalTests_Swift
 #endif
 import XCTest
