+++ ApprovalTests_SwiftTests/TestSetupTests.swift
@@ -1,13 +1,14 @@
-#if os(OSX)
-    import ApprovalTests_Swift
-#elseif os(iOS)
-    import ApprovalTests_iOS
+#if os(iOS)
+  import ApprovalTests_iOS
+#else
+  import ApprovalTests_Swift
 #endif
 
 import XCTest
 
 final class TestSetupTests: XCTestCase {
     func test() throws {
+      _ = TestSetup()
         XCTAssertTrue(TestSetup.initialized, "Test suite was not initialized")
     }
 }
