+++ Package.swift
@@ -3,23 +3,28 @@
 import PackageDescription
 
 let package = Package(
-    name: "ApprovalTests.Swift",
-    platforms: [
-        .macOS(.v10_14),
-        .iOS(.v12),
-    ],
-    products: [
-        .library(
-            name: "ApprovalTests.Swift",
-            targets: ["ApprovalTests.Swift"]
-        ),
-    ],
-    targets: [
-        .target(
-            name: "ApprovalTests.Swift",
-            dependencies: [],
-            path: "ApprovalTests.Swift",
-            exclude: ["Info.plist"]
-        ),
-    ]
+  name: "ApprovalTests.Swift",
+  platforms: [
+    .macOS(.v10_14),
+    .iOS(.v12),
+  ],
+  products: [
+    .library(
+      name: "ApprovalTests.Swift",
+      targets: ["ApprovalTests.Swift"]
+    ),
+  ],
+  targets: [
+    .target(
+      name: "ApprovalTests.Swift",
+      dependencies: [],
+      path: "ApprovalTests.Swift",
+      exclude: ["Info.plist"]
+    ),
+    .testTarget(
+      name: "ApprovalTests_SwiftTests",
+      dependencies: ["ApprovalTests.Swift"],
+      path: "ApprovalTests_SwiftTests"
+    ),
+  ]
 )
